Phase 9c: drop vestigial int32 params and upgrade rasterNativeOffset type
destroyNativeData/getSizeNativeData always received (o, 0, 0); the two
trailing params were unnamed in every implementation. Strip them to
(void*) across all 6 platform files, 5 headers, and the geoplg.cpp
dispatcher.

Driver::rasterNativeOffset promoted from int32 to ptrdiff_t — matches
the PluginOffset::value() return type and removes all the
static_cast<int32> casts that littered the platform construct lambdas.

+++ src/gl/gl3.cpp
@@ -36,7 +36,7 @@ registerPlatformPlugins(void)
 #ifdef RW_OPENGL
 	            engine->driver[PLATFORM_GL3]->defaultPipeline = makeDefaultPipeline();
 #endif
-	            engine->driver[PLATFORM_GL3]->rasterNativeOffset = static_cast<int32>(nativeRasterOffset.value());
+	            engine->driver[PLATFORM_GL3]->rasterNativeOffset = nativeRasterOffset.value();
 	            engine->driver[PLATFORM_GL3]->rasterCreate       = rasterCreate;
 	            engine->driver[PLATFORM_GL3]->rasterLock         = rasterLock;
 	            engine->driver[PLATFORM_GL3]->rasterUnlock       = rasterUnlock;
